Rename package to shardstreams
For consistency with distributor package.

We use ingesterTimeSharding name where needed to avoid clashes.
