+++ pkg/ingester/limiter.go
@@ -11,7 +11,7 @@ import (
 
 	"github.com/grafana/loki/v3/pkg/compactor/retention"
 	"github.com/grafana/loki/v3/pkg/distributor/shardstreams"
-	"github.com/grafana/loki/v3/pkg/ingester/streamsharding"
+	ingesterTimeSharding "github.com/grafana/loki/v3/pkg/ingester/shardstreams"
 	"github.com/grafana/loki/v3/pkg/validation"
 )
 
@@ -33,7 +33,7 @@ type Limits interface {
 	MaxGlobalStreamsPerUser(userID string) int
 	PerStreamRateLimit(userID string) validation.RateLimit
 	ShardStreams(userID string) shardstreams.Config
-	IngesterTimeSharding(userID string) streamsharding.Config
+	IngesterTimeSharding(userID string) ingesterTimeSharding.Config
 	IngestionPartitionsTenantShardSize(userID string) int
 
 	validation.IngestionPolicyOverrideLimits
